requirements.txt
fastapi~=0.73.0
numpy~=1.22.0
pandas~=1.3.5
sklearn~=0.0
scikit-learn~=1.0.2
python-dotenv~=0.19.2